auth/requirements.txt
fastapi>=0.115.0
stripe>=7.0.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
slowapi>=0.1.9
redis>=5.0.0
google-auth>=2.29.0
requests>=2.31.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
sqlalchemy>=2.0.0
